use GetUserWorkspace in proxy, separate get and list from space lister
