+++ pkg/proxy/handlers/spacelister_test.go
@@ -45,6 +45,9 @@ func TestSpaceLister(t *testing.T) {
 		newSignup("animelover", "anime.lover", true),
 		newSignup("carlover", "car.lover", true),
 		newSignup("racinglover", "racing.lover", false),
+		newSignup("parentspace", "parent.space", true),
+		newSignup("childspace", "child.space", true),
+		newSignup("grandchildspace", "grandchild.space", true),
 	)
 
 	// space that is not provisioned yet
@@ -81,6 +84,11 @@ func TestSpaceLister(t *testing.T) {
 		fake.NewSpace("animelover", "member-1", "animelover"),
 		fake.NewSpace("carlover", "member-1", "carlover"),
 		spaceNotProvisionedYet,
+		fake.NewSpace("parentspace", "member-1", "parentspace"),
+		fake.NewSpace("childspace", "member-1", "childspace", spacetest.WithSpecParentSpace("parentspace")),
+		fake.NewSpace("grandchildspace", "member-1", "grandchildspace", spacetest.WithSpecParentSpace("childspace")),
+		// noise space, user will have a different role here , just to make sure this is not returned anywhere in the tests
+		fake.NewSpace("otherspace", "member-1", "otherspace", spacetest.WithSpecParentSpace("otherspace")),
 
 		//spacebindings
 		fake.NewSpaceBinding("dancer-sb1", "dancelover", "dancelover", "admin"),
@@ -93,6 +101,11 @@ func TestSpaceLister(t *testing.T) {
 		spaceBindingWithSBRonDanceLover,
 		spaceBindingWithInvalidSBRName,
 		spaceBindingWithInvalidSBRNamespace,
+		fake.NewSpaceBinding("parent-sb1", "parentspace", "parentspace", "admin"),
+		fake.NewSpaceBinding("child-sb1", "childspace", "childspace", "admin"),
+		fake.NewSpaceBinding("grandchild-sb1", "grandchildspace", "grandchildspace", "admin"),
+		// noise spacebinding, just to make sure this is not returned anywhere in the tests
+		fake.NewSpaceBinding("parent-sb2", "parentspace", "otherspace", "contributor"),
 
 		//nstemplatetier
 		fake.NewBase1NSTemplateTier(),
@@ -488,6 +501,175 @@ func TestSpaceLister(t *testing.T) {
 				expectedErrCode:   500,
 				expectedWorkspace: "carlover",
 			},
+			"parent can list parentspace": {
+				username: "parent.space",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "parentspace", "admin", true,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string(nil), // this is system generated so no actions for the user
+							},
+						}),
+					),
+				},
+				expectedWorkspace: "parentspace",
+			},
+			"parent can list childspace": {
+				username: "parent.space",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "childspace", "admin", false,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "childspace",
+								Role:             "admin",
+								AvailableActions: []string(nil),
+							},
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+						}),
+					),
+				},
+				expectedWorkspace: "childspace",
+			},
+			"parent can list grandchildspace": {
+				username: "parent.space",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "grandchildspace", "admin", false,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "childspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+							{
+								MasterUserRecord: "grandchildspace",
+								Role:             "admin",
+								AvailableActions: []string(nil),
+							},
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+						}),
+					),
+				},
+				expectedWorkspace: "grandchildspace",
+			},
+			"child cannot list parentspace": {
+				username:          "child.space",
+				expectedWs:        []toolchainv1alpha1.Workspace{},
+				expectedErr:       "\"workspaces.toolchain.dev.openshift.com \\\"parentspace\\\" not found\"",
+				expectedErrCode:   404,
+				expectedWorkspace: "parentspace",
+			},
+			"child can list childspace": {
+				username:          "child.space",
+				expectedWorkspace: "childspace",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "childspace", "admin", true,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "childspace",
+								Role:             "admin",
+								AvailableActions: []string(nil),
+							},
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+						}),
+					),
+				},
+			},
+			"child can list grandchildspace": {
+				username:          "child.space",
+				expectedWorkspace: "grandchildspace",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "grandchildspace", "admin", false,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "childspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+							{
+								MasterUserRecord: "grandchildspace",
+								Role:             "admin",
+								AvailableActions: []string(nil),
+							},
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+						}),
+					),
+				},
+			},
+			"grandchild can list grandchildspace": {
+				username: "grandchild.space",
+				expectedWs: []toolchainv1alpha1.Workspace{
+					workspaceFor(t, fakeClient, "grandchildspace", "admin", true,
+						commonproxy.WithAvailableRoles([]string{
+							"admin", "viewer",
+						}),
+						commonproxy.WithBindings([]toolchainv1alpha1.Binding{
+							{
+								MasterUserRecord: "childspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+							{
+								MasterUserRecord: "grandchildspace",
+								Role:             "admin",
+								AvailableActions: []string(nil),
+							},
+							{
+								MasterUserRecord: "parentspace",
+								Role:             "admin",
+								AvailableActions: []string{"override"},
+							},
+						}),
+					),
+				},
+				expectedWorkspace: "grandchildspace",
+			},
+			"grandchild cannot list parentspace": {
+				username:          "grandchild.space",
+				expectedWorkspace: "parentspace",
+				expectedErr:       "\"workspaces.toolchain.dev.openshift.com \\\"parentspace\\\" not found\"",
+				expectedErrCode:   404,
+				expectedWs:        []toolchainv1alpha1.Workspace{},
+			},
+			"grandchild cannot list childspace": {
+				username:          "grandchild.space",
+				expectedWorkspace: "childspace",
+				expectedErr:       "\"workspaces.toolchain.dev.openshift.com \\\"childspace\\\" not found\"",
+				expectedErrCode:   404,
+				expectedWs:        []toolchainv1alpha1.Workspace{},
+			},
 		}
 
 		for k, tc := range tests {
@@ -542,143 +724,6 @@ func TestSpaceLister(t *testing.T) {
 			})
 		}
 	})
-}
-
-func TestListUserWorkspaces(t *testing.T) {
-	fakeSignupService := fake.NewSignupService(
-		newSignup("parentspace", "parent.space", true),
-		newSignup("childspace", "child.space", true),
-		newSignup("grandchildspace", "grandchild.space", true),
-	)
-
-	fakeClient := fake.InitClient(t,
-		// spaces
-		fake.NewSpace("parentspace", "member-1", "parentspace"),
-		fake.NewSpace("childspace", "member-1", "childspace", spacetest.WithSpecParentSpace("parentspace")),
-		fake.NewSpace("grandchildspace", "member-1", "grandchildspace", spacetest.WithSpecParentSpace("childspace")),
-		// noise space, user will have a different role here , just to make sure this is not returned anywhere in the tests
-		fake.NewSpace("otherspace", "member-1", "otherspace", spacetest.WithSpecParentSpace("otherspace")),
-
-		//spacebindings
-		fake.NewSpaceBinding("parent-sb1", "parentspace", "parentspace", "admin"),
-		fake.NewSpaceBinding("child-sb1", "childspace", "childspace", "admin"),
-		fake.NewSpaceBinding("grandchild-sb1", "grandchildspace", "grandchildspace", "admin"),
-		// noise spacebinding, just to make sure this is not returned anywhere in the tests
-		fake.NewSpaceBinding("parent-sb2", "parentspace", "otherspace", "contributor"),
-
-		//nstemplatetier
-		fake.NewBase1NSTemplateTier(),
-	)
-
-	// given
-	tests := map[string]struct {
-		username             string
-		expectedWs           []toolchainv1alpha1.Workspace
-		forWorkspace         string
-		overrideSignupFunc   func(ctx *gin.Context, userID, username string, checkUserSignupComplete bool) (*signup.Signup, error)
-		overrideInformerFunc func() service.InformerService
-	}{
-		"parent can list parentspace": {
-			username: "parent.space",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "parentspace", "admin", true),
-			},
-			forWorkspace: "parentspace",
-		},
-		"parent can list childspace": {
-			username: "parent.space",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "childspace", "admin", false),
-			},
-			forWorkspace: "childspace",
-		},
-		"parent can list grandchildspace": {
-			username: "parent.space",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "grandchildspace", "admin", false),
-			},
-			forWorkspace: "grandchildspace",
-		},
-		"child cannot list parentspace": {
-			username:     "child.space",
-			expectedWs:   []toolchainv1alpha1.Workspace{},
-			forWorkspace: "parentspace",
-		},
-		"child can list childspace": {
-			username:     "child.space",
-			forWorkspace: "childspace",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "childspace", "admin", true),
-			},
-		},
-		"child can list grandchildspace": {
-			username:     "child.space",
-			forWorkspace: "grandchildspace",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "grandchildspace", "admin", false),
-			},
-		},
-		"grandchild can list grandchildspace": {
-			username:     "grandchild.space",
-			forWorkspace: "grandchildspace",
-			expectedWs: []toolchainv1alpha1.Workspace{
-				workspaceFor(t, fakeClient, "grandchildspace", "admin", true),
-			},
-		},
-		"grandchild cannot list parentspace": {
-			username:     "grandchild.space",
-			forWorkspace: "parentspace",
-			expectedWs:   []toolchainv1alpha1.Workspace{},
-		},
-		"grandchild cannot list childspace": {
-			username:     "grandchild.space",
-			forWorkspace: "childspace",
-			expectedWs:   []toolchainv1alpha1.Workspace{},
-		},
-	}
-
-	for k, tc := range tests {
-		t.Run(k, func(t *testing.T) {
-			// given
-			signupProvider := fakeSignupService.GetSignupFromInformer
-			if tc.overrideSignupFunc != nil {
-				signupProvider = tc.overrideSignupFunc
-			}
-
-			informerFunc := fake.GetInformerService(fakeClient)
-			if tc.overrideInformerFunc != nil {
-				informerFunc = tc.overrideInformerFunc
-			}
-
-			proxyMetrics := metrics.NewProxyMetrics(prometheus.NewRegistry())
-
-			s := &handlers.SpaceLister{
-				GetSignupFunc:          signupProvider,
-				GetInformerServiceFunc: informerFunc,
-				ProxyMetrics:           proxyMetrics,
-			}
-
-			e := echo.New()
-			req := httptest.NewRequest(http.MethodGet, "/", strings.NewReader(""))
-			rec := httptest.NewRecorder()
-			ctx := e.NewContext(req, rec)
-			ctx.Set(rcontext.UsernameKey, tc.username)
-			ctx.Set(rcontext.RequestReceivedTime, time.Now())
-
-			// when
-			workspaceList, err := s.ListUserWorkspaces(ctx, tc.forWorkspace)
-
-			// then
-			require.NoError(t, err)
-			// list workspace case
-			require.Equal(t, len(tc.expectedWs), len(workspaceList))
-			for i := range tc.expectedWs {
-				assert.Equal(t, tc.expectedWs[i].Name, workspaceList[i].Name)
-				assert.Equal(t, tc.expectedWs[i].Status, workspaceList[i].Status)
-			}
-		})
-	}
-
 }
 
 func newSignup(signupName, username string, ready bool) fake.SignupDef {
