+++ pkg/proxy/proxy_test.go
@@ -17,17 +17,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/codeready-toolchain/registration-service/pkg/metrics"
-	"github.com/prometheus/client_golang/prometheus"
-
 	appservice "github.com/codeready-toolchain/registration-service/pkg/application/service"
 	"github.com/codeready-toolchain/registration-service/pkg/auth"
+	"github.com/codeready-toolchain/registration-service/pkg/metrics"
 	"github.com/codeready-toolchain/registration-service/pkg/proxy/access"
 	"github.com/codeready-toolchain/registration-service/pkg/proxy/handlers"
 	"github.com/codeready-toolchain/registration-service/pkg/proxy/service"
 	"github.com/codeready-toolchain/registration-service/pkg/signup"
 	"github.com/codeready-toolchain/registration-service/test"
 	"github.com/codeready-toolchain/registration-service/test/fake"
+	"github.com/prometheus/client_golang/prometheus"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
 	commoncluster "github.com/codeready-toolchain/toolchain-common/pkg/cluster"
@@ -599,6 +598,9 @@ func (s *TestProxySuite) TestProxy() {
 													}, nil
 												}
 												return nil, fmt.Errorf("proxy plugin not found")
+											}
+											inf.GetNSTemplateTierFunc = func(tier string) (*toolchainv1alpha1.NSTemplateTier, error) {
+												return fake.NewBase1NSTemplateTier(), nil
 											}
 											s.Application.MockInformerService(inf)
 											fakeApp.MemberClusterServiceMock = s.newMemberClusterServiceWithMembers(testServer.URL)
